Move fallbacks into AccordionListCell
